+++ tecrep-equipments-management-domain/src/main/java/mc/monacotelecom/tecrep/equipments/repository/HomologacionMaterialSapRepository.java
@@ -1,0 +1,7 @@
+package mc.monacotelecom.tecrep.equipments.repository;
+
+import mc.monacotelecom.tecrep.equipments.entity.HomologacionMaterialSap;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface HomologacionMaterialSapRepository extends JpaRepository<HomologacionMaterialSap, Long> {
+}