+++ tecrep-equipments-management-domain/src/main/java/mc/monacotelecom/tecrep/equipments/repository/PoAncillaryEquipmentSapRepository.java
@@ -1,0 +1,7 @@
+package mc.monacotelecom.tecrep.equipments.repository;
+
+import mc.monacotelecom.tecrep.equipments.entity.PoAncillaryEquipmentSap;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PoAncillaryEquipmentSapRepository extends JpaRepository<PoAncillaryEquipmentSap, Long> {
+}